+++ src/Typography/index.ts
@@ -2,3 +2,4 @@ export { Text } from './Text';
 export { Caption } from './Caption';
 export { Title } from './Title';
 export { Subtitle } from './Subtitle';
+export { Numeric } from './Numeric';