+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.6.2](https://github.com/samokat-oss/operations-ui/compare/v0.6.1...v0.6.2) (2020-02-07)
+
+
+### Bug Fixes
+
+* **typography:** remove unneccesary symbol ([ec06fcb](https://github.com/samokat-oss/operations-ui/commit/ec06fcb))
+
+
+
+### [0.6.1](https://github.com/samokat-oss/operations-ui/compare/v0.6.0...v0.6.1) (2020-02-06)
+
+
+### Bug Fixes
+
+* **typography:** export numeric component ([8fd6855](https://github.com/samokat-oss/operations-ui/commit/8fd6855))
+
+
+
+## [0.6.0](https://github.com/samokat-oss/operations-ui/compare/v0.5.6...v0.6.0) (2020-02-06)
+
+
+### Bug Fixes
+
+* fix version of react-input-mask ([1fb81b4](https://github.com/samokat-oss/operations-ui/commit/1fb81b4))
+
+
+### Features
+
+* **numeric component:** new numeric component ([16fcf03](https://github.com/samokat-oss/operations-ui/commit/16fcf03))
+* **typography:** bold text new tittle ([7aa0915](https://github.com/samokat-oss/operations-ui/commit/7aa0915))
+
+
+
 ### [0.5.6](https://github.com/samokat-oss/operations-ui/compare/v0.5.5...v0.5.6) (2020-02-04)